+++ dev/io.openliberty.data.internal_fat_exp/bnd.bnd
@@ -1,0 +1,36 @@
+#*******************************************************************************
+# Copyright (c) 2024 IBM Corporation and others.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License 2.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-2.0/
+# 
+# SPDX-License-Identifier: EPL-2.0
+#
+# Contributors:
+#     IBM Corporation - initial API and implementation
+#*******************************************************************************
+-include= ~../cnf/resources/bnd/bundle.props
+bVersion=1.0
+
+src: \
+  fat/src,\
+  test-applications/DataExperimentalWeb/src
+
+javac.source: 21
+javac.target: 21
+
+fat.minimum.java.level: 21
+fat.project: true
+
+-buildpath: \
+  com.ibm.ws.componenttest.2.0;version=latest,\
+  fattest.simplicity;version=latest,\
+  io.openliberty.data,\
+  io.openliberty.jakarta.annotation.2.1,\
+  io.openliberty.jakarta.cdi.4.0,\
+  io.openliberty.jakarta.data.1.0,\
+  io.openliberty.jakarta.interceptor.2.1,\
+  io.openliberty.jakarta.persistence.3.1,\
+  io.openliberty.jakarta.servlet.6.0;version=latest,\
+  io.openliberty.jakarta.transaction.2.0;version=latest
