+++ dev/io.openliberty.data.internal_fat_exp/publish/servers/io.openliberty.data.internal.fat.exp/server.xml
@@ -1,0 +1,37 @@
+<!--
+    Copyright (c) 2024 IBM Corporation and others.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License 2.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-2.0/
+    
+    SPDX-License-Identifier: EPL-2.0
+   
+    Contributors:
+        IBM Corporation - initial API and implementation
+ -->
+<server>
+
+  <featureManager>
+    <feature>componenttest-2.0</feature>
+    <feature>data-1.0</feature>
+    <feature>persistence-3.1</feature>
+    <feature>servlet-6.0</feature>
+  </featureManager>
+
+  <include location="../fatTestPorts.xml"/>
+
+  <application location="DataExperimentalApp.ear"/>
+
+  <authData id="auth1" user="auth1user" password="auth1pwd"/>
+
+  <dataSource id="DefaultDataSource" containerAuthDataRef="auth1">
+    <jdbcDriver libraryRef="DerbyLib"/>
+    <properties.derby.embedded createDatabase="create" databaseName="memory:defaultdb"/>
+  </dataSource>
+
+  <library id="DerbyLib">
+    <file name="${shared.resource.dir}/derby/derby.jar"/>
+  </library>
+
+</server>
